refactor: extract worktree session-retarget logic
Move the session execution-target retarget, rollback, reminder-state, and
git-provenance logic out of service.go into worktree_retarget.go.
